More sensible + fibs work

+++ cao-lang/src/compiler/card.rs
@@ -171,11 +171,11 @@ impl Card {
 
 #[derive(Debug, Clone, Default, Copy)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub struct IntegerNode(pub i32);
+pub struct IntegerNode(pub i64);
 
 #[derive(Debug, Clone, Default, Copy)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub struct FloatNode(pub f32);
+pub struct FloatNode(pub f64);
 
 #[derive(Debug, Clone, Default)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
